switch to 32-bit floating point for cuda compatibility and speed

+++ knn_parallel.pyx
@@ -5,8 +5,8 @@ from libc.math cimport fabs as c_fabs
 import numpy as np
 cimport numpy as np
 
-DTYPE = np.float64
-ctypedef np.float64_t DTYPE_t
+DTYPE = np.float32
+ctypedef np.float32_t DTYPE_t
 
 from cython.parallel import prange
 cimport cython
